[BACKEND] Ajustes fluxo de pacientes para notificação de vacinas.

+++ src/main/java/vetcare/api/model/dto/AnimalClienteDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AnimalClienteDTO {
 
-    Long idPet;
+    Integer idPet;
     String nomePet;
     String racaPet;
     Double pesoPet;
